#438 step card logic moved to base class
