+++ src/lib/how/rc/cards/sz-how-rc-final-entity-card.component.html
@@ -1,4 +1,4 @@
-<mat-icon class="toggle-child-expansion-icon" (click)="togglExpansion()" fontIcon="play_circle"></mat-icon>
+<mat-icon class="toggle-group-expansion-icon" (click)="toggleExpansion()" fontIcon="play_circle"></mat-icon>
 <h3>
     {{title}}<span class="datasources"> [ <span class="datasources-list">{{dataSourcesAsString}}</span> ]</span>
 </h3>
