--- src/lib/how/rc/cards/sz-how-rc-final-entity-card.component.scss
@@ -1,65 +1,0 @@
-:host {
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    flex-direction: row;
-    background-color: var(var(--sz-how-step-card-type-color-final));
-    border: var(--sz-how-step-card-border);
-    color: var(--sz-how-step-card-color);
-    margin-top: 14px;
-    border-radius: 5px;
-    padding: 3px 3px;
-    box-shadow: 0px 3px 7px 0px rgb(0 0 0 / 35%), 
-        0px 11px 12px 4px rgb(0 0 0 / 14%), 
-        11px 0px 18px 4px rgb(0 0 0 / 10%);
-
-    &.type-add {
-        background-color: var(--sz-how-step-card-type-color-add);
-    }
-    &.type-create {
-        background-color: var(--sz-how-step-card-type-color-create);
-    }
-    &.type-interim {
-        background-color: var(--sz-how-step-card-type-color-interim);
-    }
-    &.type-merge {
-        background-color: var(--sz-how-step-card-type-color-merge);
-    }
-    &.type-final {
-        background-color: var(--sz-how-step-card-type-color-final);
-    }
-
-    h3 {
-        margin: 0;
-    }
-
-    .toggle-child-expansion-icon {
-        cursor: pointer;
-        transform: rotate(0deg);
-        transition: transform 300ms;
-    }
-
-    /*&.collapsed {
-        .toggle-child-expansion-icon {
-            transform: rotate(0deg);
-        }
-    }
-    &.expanded {
-        .toggle-child-expansion-icon {
-            transform: rotate(90deg);
-        }
-    }*/
-
-    ul {
-        list-style: none;
-        li {
-
-        }
-    }
-
-    .datasources {
-        .datasources-list {
-            font-weight: normal;
-        }
-    }
-}
